Add new SocketCommunicationError enum
This adds an enum that captures both read and write errors to/from the socket.

I tried to "break" the `write_all` call by terminating the client between a read and a write (with a sleep in between), but that didn't in fact generate an error, as noted by Wgaffa@Twitch: "it is not considered an error if the entire bugger could not be written to this writer".