Add logging
Adds simple logging to the server using `env_logger`. Currently the output generated by `log_communication_error` isn't great as _logging_ output (e.g., it's not easy to automatically parse), but we decided not to get too bogged down there for now.

We needed to introduce logging in part to capture errors that happened in the futures generated by `tokio:spawn` since we can't propagate errors outside of threads.